Finalize with the styling
